pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "supportops-openenv"
version = "1.0.0"
description = "Customer support ticket triage environment"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.0",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"
[tool.setuptools]
packages = ["server"]
include-package-data = true